requirements.txt
aiohttp==3.7.4.post0
aiolimiter==1.0.0b1
async-timeout==3.0.1
attrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
cryptography==37.0.4
debugpy==1.6.3
discord.py==2.4.0
dnspython==2.2.1
dontuserepl==0.1.5
easy-pil==0.1.9
fast-colorthief==0.0.4
humanize==4.3.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
jishaku==2.5.1
motor==3.0.0
multidict==6.0.2
nest-asyncio==1.5.5
numpy==1.23.2
packaging==21.3
Pillow==8.4.0
pycparser==2.21
pydantic==1.10.2
PyJWT==2.4.0
pymongo==4.3.3
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==0.20.0
PyYAML==6.0
requests==2.28.1
six==1.16.0
typing-extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
yarl==1.8.1
